requirements.txt
bcrypt==3.1.4
Django==1.11
mysqlclient==1.3.7
PyJWT==1.6.4
tornado==4.5.2
websocket==0.2.1
websocket-client==0.48.0
